Move parMETIS implementation to core/parmetis.c
